feat: Add command output configuration
This commit adds a new configuration option `command-output` to the `wm-laravel-dem.php` config file. The value of this option is set to "This is the command output".

The `WmLaravelDemCommand` class has been modified to use the configured value as its output when executed. The command signature has also been changed from `wm-laravel-dem` to `wm-laravel-dem:fake`.

A new test file, `ExampleCommandTest.php`, has been added to test the behavior of the command. It includes two tests: one that checks if the command outputs the configured value and another that verifies if it can output a different value when the configuration is updated.

The old test file, `ExampleTest.php`, has been deleted as it is no longer needed.

Note: This commit follows semantic versioning guidelines.